# 26 sort data early