## Summary
By sorting the data early the reduce function doesn't have to search for prices with the same SKU.

## Description
Simple benchmarks running the CLI with the `time` command:
Before:
```
26000 rows: 15s
15000 rows: 6s
5000 rows: 1.9s
1000 rows: 1s
```

After:
```
26000 rows: 3s
15000 rows: 2.5s
5000 rows: 1.3s
1000 rows: 0.8s
```